chore: Don't use version pinning for Alpine packages (#42)
It turns out that it's not practically feasible to pin package versions,
because the Alpine team prunes the previous versions whenever they
publish a new distrib release (be it major, minor or patch). This
regularly breaks the build of this image and of its derived images, so
we have to get rid of version pinning. This is a shame, but no other
choice for now.